using commons-codec for file hash. Added content mapping component

+++ src/test/java/com/vektorsoft/appbox/server/test/TestUtil.java
@@ -15,38 +15,43 @@
 package com.vektorsoft.appbox.server.test;
 
 import com.vektorsoft.appbox.server.content.FileSystemContentLocatorTest;
-import com.vektorsoft.xapps.deployer.client.HashCalculator;
+import org.apache.commons.codec.digest.DigestUtils;
+import org.apache.commons.codec.digest.MessageDigestAlgorithms;
+
 import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 
 /**
- *
  * @author Vladimir Djurovic <vdjurovic@vektorsoft.com>
  */
 public final class TestUtil {
-    
-    public static final String MOCK_HASH = "f1b714039e8a09e079e86f6b436386992d2c4c05";
 
-    public static void createTestContent() throws Exception {
-	File contentDir = new File("target","content");
-	File dataDir = new File(contentDir, "content");
-	dataDir.mkdirs();
-	String hash = HashCalculator.fileHash(new File(FileSystemContentLocatorTest.class.getResource("/content/mock_content").toURI()));
-	String[] parts = new String[]{
-	    hash.substring(0, 2),
-	    hash.substring(2,4),
-	    hash.substring(4,6)
-	};
-	Path fileDir = Path.of("target", "content","content", parts[0], parts[1], parts[2]);
-	fileDir.toFile().mkdirs();
-	Path contentPath = Path.of("target", "content","content", parts[0], parts[1], parts[2], hash);
-	Files.copy(FileSystemContentLocatorTest.class.getResourceAsStream("/content/mock_content"), contentPath, StandardCopyOption.REPLACE_EXISTING);
-    }
-    
-    public static void clearTestContent() {
-	File contentDir = new File("target", "content");
-	contentDir.delete();
-    }
+	public static final String MOCK_HASH = "49b97dc6dfffb11f90b3ea378dd3bc19f5d44b7d";
+	private static final DigestUtils DIGEST = new DigestUtils(MessageDigestAlgorithms.SHA_1);
+
+	public static void createTestContent() throws Exception {
+		File contentDir = new File("target", "content");
+		File dataDir = new File(contentDir, "content");
+		dataDir.mkdirs();
+		String hash = DIGEST.digestAsHex(new File(FileSystemContentLocatorTest.class.getResource("/content/mock_content").toURI()));
+		String[] parts = new String[]{
+				hash.substring(0, 2),
+				hash.substring(2, 4),
+				hash.substring(4, 6)
+		};
+		Path fileDir = Path.of("target", "content", "content", parts[0], parts[1], parts[2]);
+		fileDir.toFile().mkdirs();
+		Path contentPath = Path.of("target", "content", "content", parts[0], parts[1], parts[2], hash);
+		if(!contentPath.toFile().exists()) {
+			Files.copy(FileSystemContentLocatorTest.class.getResourceAsStream("/content/mock_content"), contentPath, StandardCopyOption.REPLACE_EXISTING);
+		}
+
+	}
+
+	public static void clearTestContent() {
+		File contentDir = new File("target", "content");
+		contentDir.delete();
+	}
 }
